backend/requirements.txt
# Flask框架
flask>=3.0.0
flask-cors>=4.0.0

# Zep Cloud SDK
zep-cloud>=2.0.0

# OpenAI SDK（用于LLM调用）
openai>=1.0.0

# PDF处理
PyMuPDF>=1.24.0

# 环境变量
python-dotenv>=1.0.0

# 数据验证
pydantic>=2.0.0

# 文件处理
werkzeug>=3.0.0

# OASIS社交媒体模拟框架
oasis-ai>=0.1.0
camel-ai>=0.2.0

